Modificaciones
Se cambia de lugar elección accesibilidad: ahora está en homepage.
